refactor for further strategyQA evaluator
